packages/azure-investigator-core/pyproject.toml
[project]
name = "azure-investigator-core"
version = "0.1.0"
description = "Shared core for the azure-investigator skill family: auth, snapshotting, pricing, az wrapper, schemas."
readme = "../../README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Mindaugas Nakrosis", email = "mindaugasm@intelme.ai" }]
dependencies = [
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "typer>=0.12",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "platformdirs>=4.2",
]

[project.scripts]
azure-investigator = "azure_investigator_core.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_investigator_core"]
